HumanIntelligence/requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chex==0.1.82
contourpy==1.1.0
crispy-bootstrap5==0.7
cycler==0.11.0
Django==4.2.3
django-crispy-forms==2.0
etils==1.4.1
flatbuffers==23.5.26
flax==0.7.2
fonttools==4.42.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4
image==1.5.33
importlib-resources==6.0.1
jax==0.4.14
jaxlib==0.4.14
keras==2.11.0
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mediapipe==0.10.3
ml-dtypes==0.2.0
msgpack==1.0.5
nest-asyncio==1.5.7
numpy==1.25.2
oauthlib==3.2.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.3.2
packaging==20.9
Pillow==10.0.0
protobuf==3.19.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
scipy==1.11.1
six==1.16.0
sklearn==0.0.post7
sounddevice==0.4.6
sqlparse==0.4.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflowjs==3.21.0
tensorstore==0.1.41
termcolor==2.3.0
toolz==0.12.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
